front/requirements.txt.txt
streamlit
streamlit-js-eval
requests
together
faiss-cpu
numpy
sentence-transformers
rank-bm25
torch
